requirements.txt
flask==2.3.3
python-dotenv==1.0.0
scikit-learn==1.2.2
joblib==1.3.2
scipy==1.10.1
statsmodels
matplotlib
seaborn
numpy==1.26.4
tensorflow==2.19.0
pandas==2.2.2
scipy==1.10.1
scikit-learn==1.2.2

